requirements.txt
sentence-transformers==2.7.0
faiss-cpu==1.8.0
transformers==4.40.1
torch==2.2.2
streamlit==1.33.0
numpy==1.26.4
pandas==2.2.2
tiktoken==0.7.0
langchain==0.1.20
langchain-community==0.0.38
tqdm==4.66.2
pytest==8.1.1
ragas==0.1.7
